test: Only delete namespaces of successful tests (#5)
Leave the test namespace around for failed tests to make debugging easier.

+++ test/conformance/hostrules/feature.go
@@ -49,9 +49,11 @@ func InitializeScenario(ctx *godog.ScenarioContext) {
 		state = tstate.New()
 	})
 
-	ctx.AfterScenario(func(*messages.Pickle, error) {
-		// delete namespace an all the content
-		_ = kubernetes.DeleteNamespace(kubernetes.KubeClient, state.Namespace)
+	ctx.AfterScenario(func(_ *messages.Pickle, err error) {
+		// delete namespace and all the content only if there was no failure
+		if err == nil {
+			_ = kubernetes.DeleteNamespace(kubernetes.KubeClient, state.Namespace)
+		}
 	})
 }
